backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
anthropic>=0.42.0
sentence-transformers>=3.3.0
faiss-cpu>=1.7.0
PyPDF2>=3.0.0
pydantic>=2.10.0
python-dotenv>=1.0.0
numpy>=1.26.0
